Support Block2 option
This commit adds support for Block2 option as defined in RFC 7959.
Implementation works as follows:

If there's Block2 option present in CoAP request, use a block number and
a block size as indices to calculate an offset. The offset is then used
as a starting point from which HTTP response body is read. Number of
read bytes from the offset depends on the block size.

In order to prevent spamming an HTTP server with too many requests (one
for each block), a cache has been introduced to store the server's
response.  All but first CoAP requests, querying the same URL,
receive data from cache.

Max age of entries in the cache is configurable via argument. Same goes
for size of the cache.